+++ certbot/plugins/standalone_test.py
@@ -15,7 +15,7 @@ from certbot import errors
 from certbot import interfaces
 
 from certbot.tests import acme_util
-from certbot.tests import test_util
+from certbot.tests import util as test_util
 
 
 class ServerManagerTest(unittest.TestCase):
