+++ certbot/tests/display/ops_test.py
@@ -17,7 +17,7 @@ from certbot import interfaces
 
 from certbot.display import util as display_util
 
-from certbot.tests import test_util
+import certbot.tests.util as test_util
 
 
 KEY = jose.JWKRSA.load(test_util.load_vector("rsa512_key.pem"))
