+++ certbot/tests/util.py
@@ -5,6 +5,7 @@
 """
 import os
 import pkg_resources
+import shutil
 import unittest
 
 from cryptography.hazmat.backends import default_backend
@@ -14,6 +15,9 @@ import OpenSSL
 from acme import errors
 from acme import jose
 from acme import util
+
+from certbot import constants
+from certbot import storage
 
 
 def vector_path(*names):
@@ -111,3 +115,46 @@ def skip_unless(condition, reason):  # pragma: no cover
         return lambda cls: cls
     else:
         return lambda cls: None
+
+
+def make_lineage(self, testfile):
+    """Creates a lineage defined by testfile.
+
+    This creates the archive, live, and renewal directories if
+    necessary and creates a simple lineage.
+
+    :param str testfile: configuration file to base the lineage on
+
+    :returns: path to the renewal conf file for the created lineage
+    :rtype: str
+
+    """
+    lineage_name = testfile[:-len('.conf')]
+
+    conf_dir = os.path.join(
+        self.config_dir, constants.RENEWAL_CONFIGS_DIR)
+    archive_dir = os.path.join(
+        self.config_dir, constants.ARCHIVE_DIR, lineage_name)
+    live_dir = os.path.join(
+        self.config_dir, constants.LIVE_DIR, lineage_name)
+
+    for directory in (archive_dir, conf_dir, live_dir,):
+        if not os.path.exists(directory):
+            os.makedirs(directory)
+
+    sample_archive = vector_path('sample-archive')
+    for kind in os.listdir(sample_archive):
+        shutil.copyfile(os.path.join(sample_archive, kind),
+                        os.path.join(archive_dir, kind))
+
+    for kind in storage.ALL_FOUR:
+        os.symlink(os.path.join(archive_dir, '{0}1.pem'.format(kind)),
+                   os.path.join(live_dir, '{0}.pem'.format(kind)))
+
+    conf_path = os.path.join(self.config_dir, conf_dir, testfile)
+    with open(vector_path(testfile)) as src:
+        with open(conf_path, 'w') as dst:
+            dst.writelines(
+                line.replace('MAGICDIR', self.config_dir) for line in src)
+
+    return conf_path
